\name{totalLoanAmount}
\alias{totalLoanAmount}
\title{Calculates the total number of loans.}
\usage{
  totalLoanAmount(lending_club_stats)
}
\arguments{
  \item{lending_club_stats}{data.frame of lending club
  stats}
}
\value{
  the total amount of loans issues through Lending Club
}
\description{
  Calculates the total number of loans.
}

